Issue #1602
Fix 1f41da64 was reverted because it broke unix sockets with relative
path and exception messages in redis.c and library.c

+++ common.h
@@ -248,7 +248,7 @@ typedef struct fold_item {
 typedef struct {
     php_stream        *stream;
     zend_string       *host;
-    short             port;
+    int               port;
     zend_string       *auth;
     double            timeout;
     double            read_timeout;
